export schemaForUse and rename it. Export steps array in custom hook
